Refactor WebApiSyntax extensions
