+++ node/src/main/scala/coop/rchain/node/NodeRuntime.scala
@@ -945,11 +945,9 @@ object NodeRuntime {
       adminWebApi = {
         implicit val ec     = engineCell
         implicit val sp     = span
-        implicit val or     = oracle
-        implicit val bs     = blockStore
         implicit val sc     = synchronyConstraintChecker
         implicit val lfhscc = lastFinalizedHeightConstraintChecker
-        new AdminWebApiImpl[F](conf.apiServer.maxBlocksLimit, blockApiLock)
+        new AdminWebApiImpl[F](blockApiLock)
       }
     } yield (
       blockStore,
