+++ node/src/main/scala/coop/rchain/node/api/AdminWebApi.scala
@@ -1,20 +1,13 @@
 package coop.rchain.node.api
 
-import cats.effect.{Concurrent, Sync}
-import cats.syntax.all._
+import cats.effect.Concurrent
 import cats.effect.concurrent.Semaphore
-import coop.rchain.blockstorage.BlockStore
-import coop.rchain.casper.{
-  LastFinalizedHeightConstraintChecker,
-  SafetyOracle,
-  SynchronyConstraintChecker
-}
+import cats.syntax.all._
 import coop.rchain.casper.api.BlockAPI
 import coop.rchain.casper.engine.EngineCell.EngineCell
+import coop.rchain.casper.{LastFinalizedHeightConstraintChecker, SynchronyConstraintChecker}
 import coop.rchain.metrics.{Metrics, Span}
-import coop.rchain.node.api.AdminWebApi._
 import coop.rchain.shared.Log
-import monix.eval.Task
 
 trait AdminWebApi[F[_]] {
   def propose: F[String]
@@ -22,25 +15,13 @@ trait AdminWebApi[F[_]] {
 
 object AdminWebApi {
 
-  class AdminWebApiImpl[F[_]: Sync: Concurrent: EngineCell: Log: Span: SafetyOracle: BlockStore: Metrics: SynchronyConstraintChecker: LastFinalizedHeightConstraintChecker](
-      apiMaxBlocksLimit: Int,
+  class AdminWebApiImpl[F[_]: Concurrent: EngineCell: SynchronyConstraintChecker: LastFinalizedHeightConstraintChecker: Log: Span: Metrics](
       blockApiLock: Semaphore[F]
   ) extends AdminWebApi[F] {
-    import AdminWebApiSyntax._
+    import WebApiSyntax._
 
     def propose: F[String] =
       BlockAPI.createBlock[F](blockApiLock).flatMap(_.liftToBlockApiErr)
-  }
-
-  final class BlockApiException(message: String) extends Exception(message)
-
-  import AdminWebApiSyntax._
-
-  object AdminWebApiSyntax {
-    implicit final class EitherStringExt[A](val x: Either[String, A]) extends AnyVal {
-      def liftToBlockApiErr[F[_]: Sync]: F[A] =
-        x.leftMap(new BlockApiException(_)).liftTo[F]
-    }
   }
 
 }
