+++ node/src/main/scala/coop/rchain/node/api/WebApi.scala
@@ -321,16 +321,4 @@ object WebApi {
     ExploratoryDeployResponse(rhoExprs, lightBlockInfo)
   }
 
-  object WebApiSyntax {
-    implicit final class OptionExt[A](val x: Option[A]) extends AnyVal {
-      def liftToSigErr[F[_]: Sync](error: String): F[A] =
-        x.liftTo[F](new SignatureException(error))
-    }
-
-    implicit final class EitherStringExt[A](val x: Either[String, A]) extends AnyVal {
-      def liftToBlockApiErr[F[_]: Sync]: F[A] =
-        x.leftMap(new BlockApiException(_)).liftTo[F]
-    }
-  }
-
 }
